[fix] : #26 패키지 및 클래스 위치 조정, rename
